# shared/openapi.json
{
  "components": {
    "schemas": {
      "BBox": {
        "properties": {
          "h": {
            "title": "H",
            "type": "number"
          },
          "w": {
            "title": "W",
            "type": "number"
          },
          "x": {
            "title": "X",
            "type": "number"
          },
          "y": {
            "title": "Y",
            "type": "number"
          }
        },
        "required": [
          "x",
          "y",
          "w",
          "h"
        ],
        "title": "BBox",
        "type": "object"
      },
...
      "BulkUpdateRequest": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "candidate_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Ids"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "exemption_code_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exemption Code Id"
          },
          "recurrence_group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurrence Group Id"
          }
        },
        "required": [
          "action"
        ],
        "title": "BulkUpdateRequest",
        "type": "object"
      },
      "BulkUpdateResponse": {
        "properties": {
          "updated": {
            "items": {
              "$ref": "#/components/schemas/CandidateOut"
            },
            "title": "Updated",
            "type": "array"
          }
        },
        "required": [
          "updated"
        ],
        "title": "BulkUpdateResponse",
        "type": "object"
      },
      "CandidateCreate": {
        "properties": {
          "bbox": {
            "$ref": "#/components/schemas/BBox"
          },
          "exemption_code_id": {
            "title": "Exemption Code Id",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "page_no": {
            "title": "Page No",
            "type": "integer"
          }
        },
        "required": [
          "page_no",
          "bbox",
          "exemption_code_id"
        ],
        "title": "CandidateCreate",
        "type": "object"
      },
      "CandidateOut": {
        "properties": {
          "ai_justification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Justification"
          },
          "bbox": {
            "$ref": "#/components/schemas/BBox"
          },
          "confidence": {
            "title": "Confidence",
            "type": "string"
          },
          "display_text": {
            "title": "Display Text",
            "type": "string"
          },
          "exemption_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exemption Code"
          },
          "exemption_code_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exemption Code Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "page_no": {
            "title": "Page No",
            "type": "integer"
          },
          "recurrence_group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurrence Group Id"
          },
          "source_rule_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Rule Key"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "page_no",
          "bbox",
          "display_text",
          "origin",
          "confidence",
          "state"
        ],
        "title": "CandidateOut",
        "type": "object"
      },
      "CandidatePatch": {
        "properties": {
          "ai_justification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Justification"
          },
          "bbox": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BBox"
              },
              {
                "type": "null"
              }
            ]
          },
          "exemption_code_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exemption Code Id"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          }
        },
        "title": "CandidatePatch",
        "type": "object"
      },
      "CertificateVerifyResponse": {
        "properties": {
          "facts": {
            "additionalProperties": true,
            "title": "Facts",
            "type": "object"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid",
          "facts"
        ],
        "title": "CertificateVerifyResponse",
        "type": "object"
      },
...
      "DocumentOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "ocr_used": {
            "title": "Ocr Used",
            "type": "boolean"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "uploaded_by": {
            "title": "Uploaded By",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "mime_type",
          "source",
          "status",
          "ocr_used",
          "uploaded_by",
          "created_at"
        ],
        "title": "DocumentOut",
        "type": "object"
      },
...
      "ExportOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "doc_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "integrity_check": {
            "additionalProperties": true,
            "title": "Integrity Check",
            "type": "object"
          },
          "manifest_version": {
            "title": "Manifest Version",
            "type": "integer"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "doc_id",
          "type",
          "sha256",
          "manifest_version",
          "integrity_check",
          "created_at"
        ],
        "title": "ExportOut",
        "type": "object"
      },
      "ExportRequest": {
        "properties": {
          "types": {
            "default": [
              "clean_pdf",
              "exemption_log_csv",
              "certificate_pdf"
            ],
            "items": {
              "type": "string"
            },
            "title": "Types",
            "type": "array"
          }
        },
        "title": "ExportRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SearchRedactRequest": {
        "properties": {
          "exemption_code_id": {
            "title": "Exemption Code Id",
            "type": "string"
          },
          "is_pattern": {
            "default": false,
            "title": "Is Pattern",
            "type": "boolean"
          },
          "page_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page No"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "scope": {
            "default": "document",
            "title": "Scope",
            "type": "string"
          }
        },
        "required": [
          "query",
          "exemption_code_id"
        ],
        "title": "SearchRedactRequest",
        "type": "object"
      },
      "SearchRedactResponse": {
        "properties": {
          "created": {
            "items": {
              "$ref": "#/components/schemas/CandidateOut"
            },
            "title": "Created",
            "type": "array"
          }
        },
        "required": [
          "created"
        ],
        "title": "SearchRedactResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "RedactProof API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/candidates/{candidate_id}": {
      "patch": {
        "operationId": "patch_candidate_route_v1_candidates__candidate_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Org-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidatePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Candidate Route",
        "tags": [
          "review"
        ]
      }
    },
    "/v1/certificates/{certificate_id}/verify": {
      "get": {
        "description": "Deliberately public \u2014 no auth dependency. See migration 0004 for the RLS policy\nthat makes this safe (declare-the-exact-id lookup, not a listing).",
        "operationId": "verify_certificate_route_v1_certificates__certificate_id__verify_get",
        "parameters": [
          {
            "in": "path",
            "name": "certificate_id",
            "required": true,
            "schema": {
              "title": "Certificate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateVerifyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Certificate Route",
        "tags": [
          "exports"
        ]
      }
    },
...
    "/v1/documents/{doc_id}/candidates": {
      "post": {
        "operationId": "create_candidate_v1_documents__doc_id__candidates_post",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Org-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Candidate",
        "tags": [
          "review"
        ]
      }
    },
    "/v1/documents/{doc_id}/candidates:bulk": {
      "post": {
        "operationId": "bulk_update_route_v1_documents__doc_id__candidates_bulk_post",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Org-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Update Route",
        "tags": [
          "review"
        ]
      }
    },
    "/v1/documents/{doc_id}/exports": {
      "post": {
        "operationId": "export_document_v1_documents__doc_id__exports_post",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Org-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExportOut"
                  },
                  "title": "Response Export Document V1 Documents  Doc Id  Exports Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Document",
        "tags": [
          "exports"
        ]
      }
    },
...
    "/v1/documents/{doc_id}/review:complete": {
      "post": {
        "operationId": "complete_review_route_v1_documents__doc_id__review_complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Org-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Review Route",
        "tags": [
          "review"
        ]
      }
    },
    "/v1/documents/{doc_id}/search-redact": {
      "post": {
        "operationId": "search_redact_route_v1_documents__doc_id__search_redact_post",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Org-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRedactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchRedactResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Redact Route",
        "tags": [
          "review"
        ]
      }
    },
...
    "/v1/exports/{export_id}/download": {
      "get": {
        "description": "specs/04-api-spec.md: \"signed URL\" in the real S3-backed design; served directly\nhere for the same reason as page previews (app/storage/local.py) \u2014 still auth-gated\nand org-scoped via get_org_db, not public like certificate verification.",
        "operationId": "download_export_v1_exports__export_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "export_id",
            "required": true,
            "schema": {
              "title": "Export Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Org-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Export",
        "tags": [
          "exports"
        ]
      }
    },
...
  }
}